# static_root/swagger.yaml
basePath: /api
definitions:
...
  error:
    properties:
      details:
        description: more details about the error.
        type: object
      error:
        description: Title of the error
        type: string
    type: object
info:
  description: These APIs let you interact with projects by arannou
  title: Arannou backend
  version: local
paths:
  /import:
    post:
      parameters:
        - description: Json file. Must be a list of objects having a data_type attribute
          in: formData
          name: import_file
          required: true
          type: file
      responses:
        "201":
          description:
            All objects added. The responses contains the new created objects
            as it might have new properties added by back end
        "400":
          description: Bad request. Overview message and detailed informations
          schema:
            $ref: "#/definitions/error"
      summary: Bulk import a bunch of objects
...
  /upload-image/{category}:
    post:
      consumes:
        - multipart/form-data
      description: Import a picture like jpeg, gif, png...
      parameters:
        - description: Image file
          in: formData
          name: image
          required: true
          type: file
        - description: Category to put the image on
          in: path
          name: category
          required: true
          type: string
      responses:
        "200":
          description: Image imported
        "400":
          description: Bad request. Overview message and detailed informations
          schema:
            $ref: "#/definitions/error"
      summary: Import a image
...
swagger: "2.0"